refactor(architecture): extract portable component module
